// src/interface.ts
export interface TodoInterface{
    id:string;
    name:string;
    isCompleted:boolean;
}


export interface TodoFormInterface{
    todos:TodoInterface[];
    handleTodoCreate: (todo:TodoInterface)=> void;
}

export interface TodoListInterface{
    handleTodoUpdate: (event: React.ChangeEvent<HTMLInputElement>,id:string)=> void;
    handleTodoRemove: (id:string)=>void;
    handleTodoComplete: (id: string)=> void;
    handleRemoveCompleted: (isCompleted: boolean)=> void;
    todos: TodoInterface[];
}

export interface TodoItemInterface{
    handleTodoUpdate: (event: React.ChangeEvent<HTMLInputElement>,id:string)=> void;
    handleTodoRemove: (id:string)=>void;
    handleTodoComplete: (id: string)=> void;
    handleRemoveCompleted: (isCompleted: boolean)=> void;
    todos: TodoInterface;
}